Use `type=gha` cache for builds in GitHub actions (#9)
This uses the `docker/build-push-action@v6` to make use of the `gha`
type build cache.

Since we're building a multi-platform image, we need different steps for
the multi-platform build and loading the image into the local Docker
daemon: The latter step can only be run successfully when not specifying
multiple platforms.